Basic setup for the PHP Wrapper
This is a first version for the PHP wrapper. Only contains get-methods
for lists and tasks (besides authentication of course).

Furthermore I've added an ical.php file which generates a valid ical
output that can be added to mac calendar for example (agenda
subscription). The agenda items are named: "Listname - Task title -
Subtask title". They are only added if a due date is set for the tasks
(subtasks are ignored if the headtask has no duedate!).

Hopefully this is helpfull to some.

I'm planning on extending the API with these functions pretty soon:

- Create a task
- Set due date for a task
- Add notes to a task
- Delete a task
- Add list
- Delete list
- Get reminders
- Set reminder for a task

That should do for most of us that want to work with Wunderlist. There
is a folder with examples, they aren't well documented, but it should
be pretty clear. To work with the demo, change the /examples/init.php
file with your own login credentials.