Change in setup of the class and added a first draft for Wunderlist Files
I'm not sure yet if there will be an abilty to get the Wunderlist Files
that are added to tasks using the API. I've set up a quick draft
(api.files.class.php) that has only one function. An example can be
found in /examples/getFiles.php - for the moment the results are still
empty so it's not funtional yet!

I've removed the "Call()" function from both classes, adding this to
base.class.php, this gave me the abiltiy to use the same function for
both classes (Wunderlist and Wunderfiles) as they extend from this
class and carry their own protected variables.

Also the class variables have been made protected instead of private
(otherwise they won't be accesible by the Wunderbase class) and the
authtoken variable has been declared public in order to pass it to the
Wunderfiles class.